fix: textarea 크기 재조정
